Enemy and visual style overhaul
Started on new stylized cel-shaded graphics and enemies now represent health remaining with different colors, starting with red as one hit left, orange with 2, and so on. Still planning on experimenting with gameplay, not sure where to actually take this project, or how far I want to go, but I plan on keeping it simple so others that want to use it have something to use, but keep the temple feeling flexible.